projects/admin/src/app/shared/components/login/login.component.ts
import { Component } from '@angular/core';
import { AbstractControl, FormControl, FormGroup } from '@angular/forms';
import { Store } from '@ngxs/store';
import { LoginUser } from '../../../state/app.actions';

@Component({
  selector: 'angular-admin-login',
  templateUrl: './login.component.html',
  styleUrls: ['./login.component.css']
})
export class LoginComponent {
  public loginForm: FormGroup = new FormGroup({
    email: new FormControl(''),
    password: new FormControl('')
  });

  constructor(private store: Store) {}

  get email(): AbstractControl {
    return this.loginForm.get('email');
  }

  get password(): AbstractControl {
    return this.loginForm.get('password');
  }

  public login(): void {
    this.store.dispatch(new LoginUser(this.email.value, this.password.value));
  }
}
